requirements.txt
scikit-learn
pandas
numpy
pyyaml
dvc
mlflow